requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
pydantic==2.8.2
redis==5.0.8
boto3==1.35.54
typesense==0.21.0
httpx==0.27.2
aiohttp==3.10.10
beautifulsoup4==4.12.3
lxml==5.2.2
readability-lxml==0.8.1
markdownify==0.12.1
